fix(ui): unify the create and connect done pages
Extract the shared done view into account-done.svelte: both flows now
show the same drive pitch, while the connect flow finishes by closing
the popup to return to the app. When the account already has a drive,
the pitch is skipped in favor of a compact "All set!" confirmation.
