+++ ui/src/routes/connect/done/+page.svelte
@@ -6,20 +6,14 @@
 <script lang="ts">
   import { onMount } from 'svelte'
 
-  import CircleCheck from '@lucide/svelte/icons/circle-check'
-
   import { goto } from '$app/navigation'
   import { resolve } from '$app/paths'
 
-  import AppHeader from '$lib/components/app-header.svelte'
-  import Polycon from '$lib/components/polycon.svelte'
-  import { Button } from '$lib/components/ui/button'
+  import AccountDone from '$lib/components/account-done.svelte'
   import routes from '$lib/routes'
   import { accountsStore } from '$lib/stores/accounts.svelte'
   import { connectStore } from '$lib/stores/connect.svelte'
   import { sessionStore } from '$lib/stores/session.svelte'
-
-  const IDENTICON_SIZE = 80
 
   const account = $derived(
     sessionStore.currentAccountId ? accountsStore.get(sessionStore.currentAccountId) : undefined,
@@ -38,29 +32,12 @@
   }
 </script>
 
-<div class="flex min-h-svh flex-col">
-  <AppHeader />
-
-  {#if account && request}
-    <main class="flex w-full flex-1 flex-col items-center justify-center px-8 pb-24">
-      <div class="flex w-full max-w-96 flex-col items-center gap-8">
-        <div class="flex flex-col items-center gap-4">
-          <Polycon
-            value={account.id.toHex()}
-            size={IDENTICON_SIZE}
-            class="shrink-0 overflow-hidden rounded-lg"
-          />
-          <p class="text-sm font-bold">{account.name}</p>
-        </div>
-
-        <div class="flex flex-col items-center gap-2 text-center">
-          <CircleCheck class="size-5" />
-          <p class="text-sm font-bold">All set!</p>
-          <p class="text-sm">Your account is connected to {request.appName}.</p>
-        </div>
-
-        <Button class="w-full" onclick={continueToApp}>Continue to app</Button>
-      </div>
-    </main>
-  {/if}
-</div>
+{#if account && request}
+  <AccountDone
+    {account}
+    toast="Connected to {request.appName}!"
+    finishLabel="Continue to app"
+    doneMessage="Your account is connected to {request.appName}."
+    onFinish={continueToApp}
+  />
+{/if}
